+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Reporting value of `errorCode` of `SourceErrorEvent` and `PlayerErrorEvent` instead of `code` for reporting error information
 
+### Fixed
+- Missing type `StreamType` when using more recent ConvivaSDK versions
+
 ## [3.5.0] - 2024-12-13
 
 ### Fixed