+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [3.6.1]
+
+### Fixed
+- Missing version number bump
+
+## [3.6.0]
 
 ### Changed
 - Reporting value of `errorCode` of `SourceErrorEvent` and `PlayerErrorEvent` instead of `code` for reporting error information
@@ -13,7 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Missing type `StreamType` when using more recent ConvivaSDK versions
 
-## [3.5.0] - 2024-12-13
+## [3.6.1] - 2024-12-13
 
 ### Fixed
 - Build error when using Bitmovin Player version 3.78.0 or newer